SciMLOperators.AddVector — Method

L = AddVector(b; [update_func, update_func!, accepted_kwargs])
L(u) = u + b

SciMLOperators.AddVector — Method

L = AddVector(B, b; [update_func, accepted_kwargs])
L(u) = u + B*b

SciMLOperators.DEFAULT_UPDATE_FUNC — Method

DEFAULT_UPDATE_FUNC(A, u, p, t)

The default update function for AbstractSciMLOperators, a no-op that leaves the operator state unchanged.

SciMLOperators.DiagonalOperator — Method

DiagonalOperator(
    diag;
    update_func,
    update_func!,
    accepted_kwargs
)

Represents a elementwise scaling (diagonal-scaling) operation that may be applied to an AbstractVecOrMat. When diag is an AbstractVector of length N, L = DiagonalOpeator(diag, ...) can be applied to AbstractArrays with size(u, 1) == N. Each column of the u will be scaled by diag, as in LinearAlgebra.Diagonal(diag) * u.

When diag is a multidimensional array, L = DiagonalOperator(diag, ...) forms an operator of size (N, N) where N = size(diag, 1) is the leading length of diag. L then is the elementwise-scaling operation on arrays of length(u) = length(diag) with leading length size(u, 1) = N.

Its state is updated by the user-provided update_func during operator evaluation (L([v,], u, p, t)), or by calls to update_coefficients[!](L, u, p, t). Both recursively call the update_function, update_func which is assumed to have the signature

update_func(diag::AbstractVecOrMat, u, p, t; <accepted kwargs>) -> new_diag

or update_func!(diag::AbstractVecOrMat, u, p, t; <accepted kwargs>) -> [modifies diag]

The set of keyword-arguments accepted by update_func[!] must be provided to MatrixOperator via the kwarg accepted_kwargs as a tuple of Symbols. kwargs cannot be passed down to update_func[!] if accepted_kwargs are not provided.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Example

SciMLOperators.cache_operator — Function

Allocate caches for L for in-place evaluation with u-like input vectors.

SciMLOperators.iscached — Method

iscached(L)

Checks whether L has preallocated caches for inplace evaluations.

SciMLOperators.isconstant — Method

isconstant(_)

Checks if an operator's state is constant or not.

SciMLOperators.islinear — Method

islinear(_)

Checks whether operator is linear or not.

SciMLOperators.update_coefficients — Function

Update the state of L based on u, input vector, p parameter object, t, and keyword arguments. Internally, update_coefficients calls the user-provided update_func method for every component operator in L with the positional arguments (u, p, t) and keyword arguments corresponding to the symbols provided to the operator via kwarg accepted_kwargs.

This method is out-of-place, i.e. fully non-mutating and Zygote-compatible.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Example

using SciMLOperator

mat_update_func = (A, u, p, t; scale = 1.0) -> p * p' * scale * t

M = MatrixOperator(zeros(4,4); update_func = mat_update_func,
                   accepted_kwargs = (:state,))

L = M + IdentityOperator(4)

u = rand(4)
p = rand(4)
t = 1.0

L = update_coefficients(L, u, p, t; scale = 2.0)
L * u

SciMLOperators.update_coefficients! — Function

Update in-place the state of L based on u, input vector, p parameter object, t, and keyword arguments. Internally, update_coefficients! calls the user-provided mutating update_func! method for every component operator in L with the positional arguments (u, p, t) and keyword arguments corresponding to the symbols provided to the operator via kwarg accepted_kwargs.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Example

using SciMLOperator

_A = rand(4, 4)
mat_update_func! = (L, u, p, t; scale = 1.0) -> copy!(A, _A)

M = MatrixOperator(zeros(4,4); update_func! = mat_update_func!)

L = M + IdentityOperator(4)

u = rand(4)
p = rand(4)
t = 1.0

update_coefficients!(L, u, p, t)
L * u

SciMLOperators.AbstractSciMLOperator — Type

abstract type AbstractSciMLOperator{T}

Subtypes of AbstractSciMLOperator represent linear, nonlinear, time-dependent operators acting on vectors, or matrix column-vectors. A lazy operator algebra is also defined for AbstractSciMLOperators.

Interface

An AbstractSciMLOperator can be called like a function. This behaves like multiplication by the linear operator represented by the AbstractSciMLOperator. Possible signatures are

L(du, u, p, t) for in-place operator evaluation

du = L(u, p, t) for out-of-place operator evaluation

Operator evaluation methods update its coefficients with (u, p, t) information using the update_coefficients(!) method. The methods are exported and can be called as follows:

update_coefficients!(L, u, p, t) for out-of-place operator update

L = update_coefficients(L, u, p, t) for in-place operator update

SciMLOperators also overloads Base.*, LinearAlgebra.mul!, LinearAlgebra.ldiv! for operator evaluation without updating operator state. An AbstractSciMLOperator behaves like a matrix in these methods. Allocation-free methods, suffixed with a ! often need cache arrays. To precache an AbstractSciMLOperator, call the function L = cache_operator(L, input_vector).

Methods

SciMLOperators.AbstractSciMLScalarOperator — Type

abstract type AbstractSciMLScalarOperator{T} <: SciMLOperators.AbstractSciMLOperator{T}

An AbstractSciMLScalarOperator represents a linear scaling operation that may be applied to Number, AbstractVecOrMat subtypes. Addition, multiplication, division of AbstractSciMLScalarOperators is defined lazily so operator state may be updated.

SciMLOperators.AddedOperator — Type

Lazy operator addition

(A1 + A2 + A3...)u = A1*u + A2*u + A3*u ....

SciMLOperators.AddedScalarOperator — Type

struct AddedScalarOperator{T, O} <: SciMLOperators.AbstractSciMLScalarOperator{T}

Lazy addition of AbstractSciMLScalarOperators

SciMLOperators.AdjointOperator — Type

struct AdjointOperator{T, LType} <: SciMLOperators.AbstractSciMLOperator{T}

SciMLOperators.AffineOperator — Type

L = AffineOperator(A, B, b; [update_func, update_func!, accepted_kwargs])
L(u) = A*u + B*b

Represents a time-dependent affine operator. The update function is called by update_coefficients! and is assumed to have the following signature:

update_func(b::AbstractArray,u,p,t; <accepted kwargs>) -> [modifies b]

SciMLOperators.BatchedDiagonalOperator — Type

BatchedDiagonalOperator(diag; update_func, update_func!, accepted_kwargs)

Represents a time-dependent elementwise scaling (diagonal-scaling) operation. Acts on AbstractArrays of the same size as diag. The update function is called by update_coefficients! and is assumed to have the following signature:

update_func(diag::AbstractArray, u, p, t; <accepted kwarg fields>) -> [modifies diag]

SciMLOperators.ComposedOperator — Type

Lazy operator composition

∘(A, B, C)(u) = A(B(C(u)))

ops = (A, B, C)
cache = (B*C*u , C*u)

SciMLOperators.ComposedScalarOperator — Type

struct ComposedScalarOperator{T, O} <: SciMLOperators.AbstractSciMLScalarOperator{T}

Lazy multiplication of AbstractSciMLScalarOperators

SciMLOperators.FunctionOperator — Type

Matrix free operators (given by a function)

op: Function with signature op(u, p, t) and (if isinplace) op(du, u, p, t)

op_adjoint: Adjoint operator

op_inverse: Inverse operator

op_adjoint_inverse: Adjoint inverse operator

traits: Traits

p: Parameters

t: Time

cache: Cache

SciMLOperators.FunctionOperator — Method

FunctionOperator(op, input)
FunctionOperator(
    op,
    input,
    output;
    isinplace,
    outofplace,
    isconstant,
    has_mul5,
    cache,
    T,
    op_adjoint,
    op_inverse,
    op_adjoint_inverse,
    p,
    t,
    accepted_kwargs,
    ifcache,
    islinear,
    opnorm,
    issymmetric,
    ishermitian,
    isposdef
)

SciMLOperators.IdentityOperator — Type

struct IdentityOperator <: SciMLOperators.AbstractSciMLOperator{Bool}

Operator representing the identity function id(u) = u

SciMLOperators.InvertedOperator — Type

Lazy Operator Inverse

SciMLOperators.InvertedScalarOperator — Type

struct InvertedScalarOperator{T, λType} <: SciMLOperators.AbstractSciMLScalarOperator{T}

Lazy inverse of AbstractSciMLScalarOperators

SciMLOperators.InvertibleOperator — Type

InvertibleOperator(L, F)

Stores an operator and its factorization (or inverse operator). Supports left division and ldiv! via F, and operator application via L.

SciMLOperators.MatrixOperator — Type

Represents a linear operator given by an AbstractMatrix that may be applied to an AbstractVecOrMat. Its state is updated by the user-provided update_func during operator evaluation (L([v,], u, p, t)), or by calls to update_coefficients[!](L, u, p, t). Both recursively call the update_function, update_func which is assumed to have the signature

update_func(A::AbstractMatrix, u, p, t; <accepted kwargs>) -> newA

or update_func!(A::AbstractMatrix, u ,p , t; <accepted kwargs>) -> [modifies A]

The set of keyword-arguments accepted by update_func[!] must be provided to MatrixOperator via the kwarg accepted_kwargs as a tuple of Symbols. kwargs cannot be passed down to update_func[!] if accepted_kwargs are not provided.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Interface

Lazy matrix algebra is defined for AbstractSciMLOperators. The Interface supports lazy addition, subtraction, multiplication, inversion, adjoints, transposes.

Example

u = rand(4)
p = rand(N, N)
t = rand()

mat_update = (A, u, p, t; scale = 0.0) -> t * p
M = MatrixOperator(0.0; update_func = mat_update; accepted_kwargs = (:scale,))

L = M * M + 3I
L = cache_operator(M, u)

# update L and evaluate
v = L(u, p, t; scale = 1.0)

v = zero(4)
u = rand(4)
p = nothing
t = rand()

mat_update! = (A, u, p, t; scale = 0.0) -> (copy!(A, p); lmul!(t, A))
M = MatrixOperator(0.0; update_func! = val_update!; accepted_kwargs = (:scale,))
L = M * M + 3I

# update L in-place and evaluate
L(v, u, p, t; scale = 1.0)

SciMLOperators.NoKwargFilter — Type

This type indicates to preprocess_update_func to not to filter keyword arguments. Required in implementation of lazy Base.adjoint, Base.conj, Base.transpose.

SciMLOperators.NullOperator — Type

struct NullOperator <: SciMLOperators.AbstractSciMLOperator{Bool}

Operator representing the null function n(u) = 0 * u

SciMLOperators.ScalarOperator — Method

ScalarOperator(val; update_func, accepted_kwargs)

Represents a linear scaling operator that may be applied to a Number, or an AbstractArray subtype. Its state is updated by the user-provided update_func during operator evaluation (L([v,] u, p, t)), or by calls to update_coefficients[!]. Both recursively call the update function, update_func which is assumed to have the signautre:

update_func(oldval::Number, u, p, t; <accepted kwargs>) -> newval

The set of keyword-arguments accepted by update_func must be provided to ScalarOperator via the kwarg accepted_kwargs as a tuple of Symbols. kwargs cannot be passed down to update_func if accepted_kwargs are not provided.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Interface

Lazy scalar algebra is defined for AbstractSciMLScalarOperators. The interface supports lazy addition, subtraction, multiplication and division.

Example

v = zero(4)
u = rand(4)
p = nothing
t = 0.0

val_update = (a, u, p, t; scale = 0.0) -> copy(scale)
α = ScalarOperator(0.0; update_func = val_update; accepted_kwargs = (:scale,))
β = 2 * α + 3 / α

# update L out-of-place, and evaluate
β(u, p, t; scale = 1.0)

# update L in-place and evaluate
β(v, u, p, t; scale = 1.0)

SciMLOperators.ScaledOperator — Type

struct ScaledOperator{T, λType, LType} <: SciMLOperators.AbstractSciMLOperator{T}

ScaledOperator

(λ L)*(u) = λ * L(u)

SciMLOperators.TensorProductOperator — Type

Lazy Tensor Product Operator

TensorProductOperator(A, B) = A ⊗ B

(A ⊗ B)(u) = vec(B * U * transpose(A))

where U is a lazy representation of the vector u as
a matrix with the appropriate size.

SciMLOperators.TransposedOperator — Type

struct TransposedOperator{T, LType} <: SciMLOperators.AbstractSciMLOperator{T}