Premade SciMLOperators

Direct Operator Definitions

Missing docstring.

Missing docstring for ScalarOperator.IdentityOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.NullOperator. Check Documenter's build log for details.

SciMLOperators.ScalarOperator — Type

ScalarOperator(val; update_func, accepted_kwargs)

Represents a linear scaling operator that may be applied to a Number, or an AbstractArray subtype. Its state is updated by the user-provided update_func during operator evaluation (L([v,] u, p, t)), or by calls to update_coefficients[!]. Both recursively call the update function, update_func which is assumed to have the signautre:

update_func(oldval::Number, u, p, t; <accepted kwargs>) -> newval

The set of keyword-arguments accepted by update_func must be provided to ScalarOperator via the kwarg accepted_kwargs as a tuple of Symbols. kwargs cannot be passed down to update_func if accepted_kwargs are not provided.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Interface

Lazy scalar algebra is defined for AbstractSciMLScalarOperators. The interface supports lazy addition, subtraction, multiplication and division.

Example

v = zero(4)
u = rand(4)
p = nothing
t = 0.0

val_update = (a, u, p, t; scale = 0.0) -> copy(scale)
α = ScalarOperator(0.0; update_func = val_update; accepted_kwargs = (:scale,))
β = 2 * α + 3 / α

# update L out-of-place, and evaluate
β(u, p, t; scale = 1.0)

# update L in-place and evaluate
β(v, u, p, t; scale = 1.0)

Missing docstring.

Missing docstring for MatrixOperator. Check Documenter's build log for details.

SciMLOperators.DiagonalOperator — Function

DiagonalOperator(
    diag;
    update_func,
    update_func!,
    accepted_kwargs
)

Represents a elementwise scaling (diagonal-scaling) operation that may be applied to an AbstractVecOrMat. When diag is an AbstractVector of length N, L = DiagonalOpeator(diag, ...) can be applied to AbstractArrays with size(u, 1) == N. Each column of the u will be scaled by diag, as in LinearAlgebra.Diagonal(diag) * u.

When diag is a multidimensional array, L = DiagonalOperator(diag, ...) forms an operator of size (N, N) where N = size(diag, 1) is the leading length of diag. L then is the elementwise-scaling operation on arrays of length(u) = length(diag) with leading length size(u, 1) = N.

Its state is updated by the user-provided update_func during operator evaluation (L([v,], u, p, t)), or by calls to update_coefficients[!](L, u, p, t). Both recursively call the update_function, update_func which is assumed to have the signature

update_func(diag::AbstractVecOrMat, u, p, t; <accepted kwargs>) -> new_diag

or update_func!(diag::AbstractVecOrMat, u, p, t; <accepted kwargs>) -> [modifies diag]

The set of keyword-arguments accepted by update_func[!] must be provided to MatrixOperator via the kwarg accepted_kwargs as a tuple of Symbols. kwargs cannot be passed down to update_func[!] if accepted_kwargs are not provided.

Warning

The user-provided update_func[!] must not use u in its computation. Positional argument (u, p, t) to update_func[!] are passed down by update_coefficients[!](L, u, p, t), where u is the input-vector to the composite AbstractSciMLOperator. For that reason, the values of u, or even shape, may not correspond to the input expected by update_func[!]. If an operator's state depends on its input vector, then it is, by definition, a nonlinear operator. We recommend sticking such nonlinearities in FunctionOperator. This topic is further discussed in (this issue)[https://github.com/SciML/SciMLOperators.jl/issues/159].

Example

Missing docstring.

Missing docstring for AffineOperator. Check Documenter's build log for details.

SciMLOperators.AddVector — Function

L = AddVector(b; [update_func, update_func!, accepted_kwargs])
L(u) = u + b

L = AddVector(B, b; [update_func, accepted_kwargs])
L(u) = u + B*b

Missing docstring.

Missing docstring for FunctionOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TensorProductOperator. Check Documenter's build log for details.

Lazy Scalar Operator Combination

Missing docstring.

Missing docstring for SciMLOperators.AddedScalarOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.ComposedScalarOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.InvertedScalarOperator. Check Documenter's build log for details.

Lazy Operator Combination

Missing docstring.

Missing docstring for SciMLOperators.ScaledOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.AddedOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.ComposedOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.InvertedOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.InvertibleOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.AdjointOperator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SciMLOperators.TransposedOperator. Check Documenter's build log for details.